+++ test/unit/operations/aggregate.test.js
@@ -10,16 +10,16 @@ describe('AggregateOperation', function () {
     context('when out is in the options', function () {
       const operation = new AggregateOperation(db, [], { out: 'test', dbName: db });
 
-      it('sets trySecondaryWrite to true', function () {
-        expect(operation.trySecondaryWrite).to.be.true;
+      it('sets hasWriteStage to true', function () {
+        expect(operation.hasWriteStage).to.be.true;
       });
     });
 
     context('when $out is the last stage', function () {
       const operation = new AggregateOperation(db, [{ $out: 'test' }], { dbName: db });
 
-      it('sets trySecondaryWrite to true', function () {
-        expect(operation.trySecondaryWrite).to.be.true;
+      it('sets hasWriteStage to true', function () {
+        expect(operation.hasWriteStage).to.be.true;
       });
     });
 
@@ -28,16 +28,16 @@ describe('AggregateOperation', function () {
         dbName: db
       });
 
-      it('sets trySecondaryWrite to false', function () {
-        expect(operation.trySecondaryWrite).to.be.false;
+      it('sets hasWriteStage to false', function () {
+        expect(operation.hasWriteStage).to.be.false;
       });
     });
 
     context('when $merge is the last stage', function () {
       const operation = new AggregateOperation(db, [{ $merge: { into: 'test' } }], { dbName: db });
 
-      it('sets trySecondaryWrite to true', function () {
-        expect(operation.trySecondaryWrite).to.be.true;
+      it('sets hasWriteStage to true', function () {
+        expect(operation.hasWriteStage).to.be.true;
       });
     });
 
@@ -48,24 +48,24 @@ describe('AggregateOperation', function () {
         { dbName: db }
       );
 
-      it('sets trySecondaryWrite to false', function () {
-        expect(operation.trySecondaryWrite).to.be.false;
+      it('sets hasWriteStage to false', function () {
+        expect(operation.hasWriteStage).to.be.false;
       });
     });
 
     context('when no writable stages in empty pipeline', function () {
       const operation = new AggregateOperation(db, [], { dbName: db });
 
-      it('sets trySecondaryWrite to false', function () {
-        expect(operation.trySecondaryWrite).to.be.false;
+      it('sets hasWriteStage to false', function () {
+        expect(operation.hasWriteStage).to.be.false;
       });
     });
 
     context('when no writable stages', function () {
       const operation = new AggregateOperation(db, [{ $project: { name: 1 } }], { dbName: db });
 
-      it('sets trySecondaryWrite to false', function () {
-        expect(operation.trySecondaryWrite).to.be.false;
+      it('sets hasWriteStage to false', function () {
+        expect(operation.hasWriteStage).to.be.false;
       });
     });
   });
